+++ frontend/src/routes/channels/id/VideoCard.tsx
@@ -1,8 +1,9 @@
-import {Card, Col, Ratio} from "react-bootstrap";
+import {Card, Col, Ratio, Row} from "react-bootstrap";
 import {VideoWithChannelIdDto} from "../../../model/VideoDto";
 import {Link} from "react-router";
 import {ArrowUpLeftCircleFill} from "react-bootstrap-icons";
 import {Date} from "../../../components/Date";
+import {CommentCounts} from "../../../components/comments/CommentCounts";
 
 type VideoCardProps = {
     video: VideoWithChannelIdDto
@@ -23,6 +24,17 @@ export const VideoCard = ({video}: VideoCardProps) => (
                     </Link>
                 </Card.Subtitle>
             </Card.Body>
+            <Card.Footer>
+                <Row>
+                    <CommentCounts
+                        comments={video.commentCount}
+                        commentsDisabled={video.commentsDisabled}
+                        replies={video.replyCount}
+                        totalReplies={video.totalReplyCount}
+                        showTotalReplies={false}
+                    />
+                </Row>
+            </Card.Footer>
             <Card.Footer className={"d-flex align-items-center"}>
                 <ArrowUpLeftCircleFill className={"me-2"}/>
                 <Date date={video.publishedAt}/>