+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Make "more replies..." item a link that fetches replies to the comment. (mddburgess/winston#8)
 - Clipboard icon on the channel page now copies the channel link to the clipboard. (mddburgess/winston#22)
+- Video card now shows the fetched comment and reply counts. (mddburgess/#20)
 
 ## [1.1.0] — 2025-04-05